+++ app/Filament/Clusters/Sgr.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Clusters;
+
+use Filament\Clusters\Cluster;
+
+class Sgr extends Cluster
+{
+    protected static ?string $navigationIcon = 'heroicon-o-rocket-launch';
+
+    protected static ?string $title = 'SGR';
+}
