+++ app/Filament/Clusters/Sgr/Resources/CategoryResource/Pages/EditCategory.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Filament\Clusters\Sgr\Resources\CategoryResource\Pages;
+
+use App\Filament\Clusters\Sgr\Resources\CategoryResource;
+use Filament\Actions;
+use Filament\Resources\Pages\EditRecord;
+
+class EditCategory extends EditRecord
+{
+    protected static string $resource = CategoryResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\ViewAction::make(),
+            Actions\DeleteAction::make(),
+        ];
+    }
+}
