+++ app/Filament/Clusters/Sgr/Resources/EventResource.php
@@ -1,0 +1,123 @@
+<?php
+
+namespace App\Filament\Clusters\Sgr\Resources;
+
+use App\Filament\Clusters\Sgr;
+use App\Filament\Clusters\Sgr\Resources\EventResource\Pages;
+use App\Filament\Clusters\Sgr\Resources\EventResource\RelationManagers;
+use App\Models\Sgr\Event;
+use Filament\Forms;
+use Filament\Forms\Form;
+use Filament\Pages\SubNavigationPosition;
+use Filament\Resources\Resource;
+use Filament\Tables;
+use Filament\Tables\Table;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\SoftDeletingScope;
+
+class EventResource extends Resource
+{
+    protected static ?string $model = Event::class;
+
+    protected static ?string $navigationIcon = 'heroicon-o-rectangle-stack';
+
+    protected static ?string $cluster = Sgr::class;
+
+    protected static SubNavigationPosition $subNavigationPosition = SubNavigationPosition::Top;
+
+    protected static ?int $navigationSort = 4;
+
+    public static function form(Form $form): Form
+    {
+        return $form
+            ->schema([
+                Forms\Components\Select::make('requirement_id')
+                    ->relationship('requirement', 'id')
+                    ->required(),
+                Forms\Components\Textarea::make('body')
+                    ->required()
+                    ->columnSpanFull(),
+                Forms\Components\Select::make('event_type_id')
+                    ->relationship('eventType', 'name')
+                    ->required(),
+                Forms\Components\DateTimePicker::make('limit_at'),
+                Forms\Components\Select::make('creator_id')
+                    ->relationship('creator', 'name')
+                    ->required(),
+                Forms\Components\Select::make('creator_ou_id')
+                    ->relationship('creatorOu', 'name')
+                    ->required(),
+                Forms\Components\Select::make('to_user_id')
+                    ->relationship('toUser', 'name'),
+                Forms\Components\Select::make('to_ou_id')
+                    ->relationship('toOu', 'name'),
+            ]);
+    }
+
+    public static function table(Table $table): Table
+    {
+        return $table
+            ->columns([
+                Tables\Columns\TextColumn::make('requirement.id')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('eventType.name')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('limit_at')
+                    ->dateTime()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('creator.name')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('creatorOu.name')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('toUser.name')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('toOu.name')
+                    ->numeric()
+                    ->sortable(),
+                Tables\Columns\TextColumn::make('created_at')
+                    ->dateTime()
+                    ->sortable()
+                    ->toggleable(isToggledHiddenByDefault: true),
+                Tables\Columns\TextColumn::make('updated_at')
+                    ->dateTime()
+                    ->sortable()
+                    ->toggleable(isToggledHiddenByDefault: true),
+                Tables\Columns\TextColumn::make('deleted_at')
+                    ->dateTime()
+                    ->sortable()
+                    ->toggleable(isToggledHiddenByDefault: true),
+            ])
+            ->filters([
+                //
+            ])
+            ->actions([
+                Tables\Actions\EditAction::make(),
+            ])
+            ->bulkActions([
+                Tables\Actions\BulkActionGroup::make([
+                    Tables\Actions\DeleteBulkAction::make(),
+                ]),
+            ]);
+    }
+
+    public static function getRelations(): array
+    {
+        return [
+            //
+        ];
+    }
+
+    public static function getPages(): array
+    {
+        return [
+            'index' => Pages\ListEvents::route('/'),
+            'create' => Pages\CreateEvent::route('/create'),
+            'edit' => Pages\EditEvent::route('/{record}/edit'),
+        ];
+    }
+}
