+++ app/Filament/Clusters/Sgr/Resources/EventResource/Pages/CreateEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Clusters\Sgr\Resources\EventResource\Pages;
+
+use App\Filament\Clusters\Sgr\Resources\EventResource;
+use Filament\Actions;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateEvent extends CreateRecord
+{
+    protected static string $resource = EventResource::class;
+}
