+++ app/Filament/Clusters/Sgr/Resources/EventResource/Pages/EditEvent.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Clusters\Sgr\Resources\EventResource\Pages;
+
+use App\Filament\Clusters\Sgr\Resources\EventResource;
+use Filament\Actions;
+use Filament\Resources\Pages\EditRecord;
+
+class EditEvent extends EditRecord
+{
+    protected static string $resource = EventResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [
+            Actions\DeleteAction::make(),
+        ];
+    }
+}
