+++ app/Filament/Clusters/Sgr/Resources/EventTypeResource/Pages/CreateEventType.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Filament\Clusters\Sgr\Resources\EventTypeResource\Pages;
+
+use App\Filament\Clusters\Sgr\Resources\EventTypeResource;
+use Filament\Actions;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateEventType extends CreateRecord
+{
+    protected static string $resource = EventTypeResource::class;
+}
