--- app/Filament/Resources/Sgr/CategoryResource/Pages/CreateCategory.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\CategoryResource\Pages;
-
-use App\Filament\Resources\Sgr\CategoryResource;
-use Filament\Actions;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateCategory extends CreateRecord
-{
-    protected static string $resource = CategoryResource::class;
-
-    protected function mutateFormDataBeforeCreate(array $data): array
-    {
-        $data['organizational_unit_id'] = auth()->user()->organizational_unit_id;
-
-        return $data;
-    }
-}
