--- app/Filament/Resources/Sgr/CategoryResource/Pages/EditCategory.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\CategoryResource\Pages;
-
-use App\Filament\Resources\Sgr\CategoryResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditCategory extends EditRecord
-{
-    protected static string $resource = CategoryResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
