--- app/Filament/Resources/Sgr/CategoryResource/Pages/ListCategories.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\CategoryResource\Pages;
-
-use App\Filament\Resources\Sgr\CategoryResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListCategories extends ListRecords
-{
-    protected static string $resource = CategoryResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
