--- app/Filament/Resources/Sgr/CategoryResource/Pages/ViewCategory.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\CategoryResource\Pages;
-
-use App\Filament\Resources\Sgr\CategoryResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ViewRecord;
-
-class ViewCategory extends ViewRecord
-{
-    protected static string $resource = CategoryResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\EditAction::make(),
-        ];
-    }
-}
