--- app/Filament/Resources/Sgr/EventResource/Pages/CreateEvent.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\EventResource\Pages;
-
-use App\Filament\Resources\Sgr\EventResource;
-use Filament\Actions;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateEvent extends CreateRecord
-{
-    protected static string $resource = EventResource::class;
-}
