--- app/Filament/Resources/Sgr/EventResource/Pages/EditEvent.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\EventResource\Pages;
-
-use App\Filament\Resources\Sgr\EventResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditEvent extends EditRecord
-{
-    protected static string $resource = EventResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
