--- app/Filament/Resources/Sgr/EventResource/Pages/ListEvents.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\EventResource\Pages;
-
-use App\Filament\Resources\Sgr\EventResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListEvents extends ListRecords
-{
-    protected static string $resource = EventResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
