--- app/Filament/Resources/Sgr/EventTypeResource/Pages/CreateEventType.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\EventTypeResource\Pages;
-
-use App\Filament\Resources\Sgr\EventTypeResource;
-use Filament\Actions;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateEventType extends CreateRecord
-{
-    protected static string $resource = EventTypeResource::class;
-}
