--- app/Filament/Resources/Sgr/EventTypeResource/Pages/EditEventType.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\EventTypeResource\Pages;
-
-use App\Filament\Resources\Sgr\EventTypeResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditEventType extends EditRecord
-{
-    protected static string $resource = EventTypeResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
