--- app/Filament/Resources/Sgr/LabelResource/Pages/EditLabel.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\LabelResource\Pages;
-
-use App\Filament\Resources\Sgr\LabelResource;
-use Filament\Actions;
-use Filament\Resources\Pages\EditRecord;
-
-class EditLabel extends EditRecord
-{
-    protected static string $resource = LabelResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\ViewAction::make(),
-            Actions\DeleteAction::make(),
-        ];
-    }
-}
