--- app/Filament/Resources/Sgr/RequirementResource/Pages/ListRequirements.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Sgr\RequirementResource\Pages;
-
-use App\Filament\Resources\Sgr\RequirementResource;
-use Filament\Actions;
-use Filament\Resources\Pages\ListRecords;
-
-class ListRequirements extends ListRecords
-{
-    protected static string $resource = RequirementResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            Actions\CreateAction::make(),
-        ];
-    }
-}
